add a few more styles for output
